requirements.txt
flask
python-dotenv
pyodbc
flask-cors